+++ src/components/cards/inventoryCard.js
@@ -1,16 +1,6 @@
 'use client';
 
-import { useEffect } from "react";
-
 export default function InventoryCard({elementName, elementId, elementQuantity}){
-  function updateModal() {
-    const modal = document.querySelector('[data-modal]');
-    const modalMaterialElement = modal.querySelector('[data-modal-material]');
-    const modalMaterialIdElement = modal.querySelector('[data-modal-material-id]');
-
-    modalMaterialElement.textContent = elementName;
-    modalMaterialIdElement.value = elementId;
-  }
   return (
     <div className="card p-1">
       <div className="title-card d-flex justify-content-start align-items-center">
@@ -21,7 +11,7 @@ export default function InventoryCard({elementName, elementId, elementQuantity})
       </div>
       <p>
         Cantidad: {elementQuantity}
-        <button data-bs-toggle="modal" data-bs-target="#updateModal" onClick={updateModal}>Modificar inventario</button>
+        <button data-open-update-modal data-element-name={elementName} data-element-id={elementId}>Modificar inventario</button>
       </p>
     </div>
   )
